#205. Moved logic for token from user controller to one place with creation of axios instance.
