+++ src/api/util/axiosInstance.js
@@ -1,12 +1,37 @@
 import axios from 'axios';
 
 import config from './config';
-import {JWT_STORAGE_NAME} from "../../components/auth/user";
+import { JWT_STORAGE_NAME } from "../../components/auth/user";
+import { Base64 } from "js-base64";
+
+// Checking if there is a valid auth token
+const hasActiveToken = () => {
+  const jwt = sessionStorage.getItem(JWT_STORAGE_NAME);
+  if (jwt) {
+    const user = JSON.parse(Base64.decode(jwt.split('.')[1]));
+    // is the token still valid - if not then delete it. This of course is only to ensure the client knows that the token has expired. any authenticated requests would fail anyhow
+    return new Date(user.exp * 1000).toISOString() >= new Date().toISOString();
+  }
+  return false;
+};
 
 // Creating axios custom instance with global base URL
 const instance = axios.create({
   baseURL: config.dataApi
 });
+
+// On startup load user from storage. Use sessionstorage for the session, but save in local storage if user choose to be remembered
+const localJwt = localStorage.getItem(JWT_STORAGE_NAME);
+if (localJwt) {
+  sessionStorage.setItem(JWT_STORAGE_NAME, localJwt);
+  if (hasActiveToken()) {
+    instance.defaults.headers.common['Authorization'] = `Bearer ${localJwt}`;
+  } else {
+    localStorage.removeItem(JWT_STORAGE_NAME);
+    sessionStorage.removeItem(JWT_STORAGE_NAME);
+    instance.defaults.headers.common['Authorization'] = '';
+  }
+}
 
 // Add a request interceptor
 instance.interceptors.response.use(
